restore FastCGI, uWSGI, and SCGI (closes #11)
This is included in all build, including Alpine as the image size
difference is only 327k and it is a common use case.

According to agentzh:
"I think we should include fastcgi and uwsgi modules by default since many
people use OpenResty as a dynamic web gateway instead of a self-contained
app server. Such hybrid setup is actually a typical use case :)"