Add custom build control as --build-args to built-from-source flavors (#79)
As per https://github.com/openresty/docker-openresty/issues/79 this commit adds four new build
variables to allow for a docker image that can be customized with additional modules.

These were applied to the built-from-source flavors: `alpine`, `xenial`, and `bionic`

Per our requirement we needed to download a module from a git repo (git clone - which required
installing a package), then add it to the ./configure (already supported by
`RESTY_CONFIG_OPTIONS_MORE` argument), and for the sakes of cleanliness perform some cleanup after
the configure as well as removing the installed package.

This is accomplished with the following `--build-args`:

 * `RESTY_ADD_PACKAGE_BUILDDEPS` - packages added to this argument are added for the build process
    and then removed after the build automatically. example: "git" or "git wget"

 * `RESTY_ADD_PACKAGE_RUNDEPS` - packages added to this argument are run-time dependencies and are
   not removed

 * `RESTY_EVAL_PRE_CONFIGURE` - any command here will be executed before ./configure (it executes
   inside the /tmp directory where other packages are also downloaded).  example: "git clone git://github.com/author/module-name"

 * `RESTY_EVAL_POST_MAKE` - any command here will execute after make install for openresty.  Often
   used for cleanup, example: `rm -rf module-name` (to remove the directory cloned)